feat: adding mail verification in client side
Adding pages to mail verification, two pages to be more
exactly. A form to make the account registrations, after that,
the backend sends a e-mail with the code to user email registered.
Then, the page is redirect to another form that receives the code,
after this, the account is verified and the user can make the login